pyproject.toml
[tool.poetry]
name = "aiogram-template"
version = "0.1.0"
description = ""
authors = ["stupidcabbage <vanya@gutche.ru>"]
packages = [{include = "telegram_bot"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
alembic = "^1.13.1"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
